validation auth in backend was completed 🔓

+++ pages/api/carrousel.js
@@ -2,46 +2,62 @@
 // Next.js API route support: https://nextjs.org/docs/api-routes/introduction
 
 import { PrismaClient } from '@prisma/client'
+import { verify } from 'jsonwebtoken'
+import dotenv from 'dotenv'
+
+dotenv.config()
 
 const prisma = new PrismaClient()
 
 export default async function handler(req, res) {
-  switch (req.method) {
-    case 'POST': {
-      const savedCarrousel = await prisma.carrousel.create({
-        data: JSON.parse(req.body)
-      })
+  const authToken = req.cookies.authToken
 
-      res.status(201).json(savedCarrousel)
-      break
-    }
+  if (authToken) {
+    const user = verify(authToken, process.env.JWT_SECRET)
 
-    case 'PUT': {
-      const { id } = req.query
+    if (user.role === 'admin' || user.role === 'editor') {
+      switch (req.method) {
+        case 'POST': {
+          const savedCarrousel = await prisma.carrousel.create({
+            data: JSON.parse(req.body)
+          })
 
-      const carrouselData = JSON.parse(req.body)
-
-      const response = await prisma.carrousel.update({
-        data: carrouselData,
-        where: {
-          id
+          res.status(201).json(savedCarrousel)
+          break
         }
-      })
-      res.status(200).json(response)
-      break
-    }
 
-    case 'DELETE': {
-      const { id } = req.query
+        case 'PUT': {
+          const { id } = req.query
 
-      const deletedCarrousel = await prisma.carrousel.delete({
-        where: {
-          id
+          const carrouselData = JSON.parse(req.body)
+
+          const response = await prisma.carrousel.update({
+            data: carrouselData,
+            where: {
+              id
+            }
+          })
+          res.status(200).json(response)
+          break
         }
-      })
 
-      res.status(200).json(deletedCarrousel)
-      break
+        case 'DELETE': {
+          const { id } = req.query
+
+          const deletedCarrousel = await prisma.carrousel.delete({
+            where: {
+              id
+            }
+          })
+
+          res.status(200).json(deletedCarrousel)
+          break
+        }
+      }
+    } else {
+      res.status(401).json({ error: 'You have not enough permissions' })
     }
+  } else {
+    res.status(401).json({ error: 'You are not authenticate' })
   }
 }
